requirements.txt
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.3
torchvision==0.8.0
pillow==8.1.2
djitellopy==1.5
dlib==19.21.99
setuptools==49.6.0
pygame==2.0.1
flask==1.1.2
requests==2.25.1
werkzeug==1.0.1